Supports Scala Controller.
We would like to convert all controllers to Scala eventually.
This is the first patch to support Scala controller.

As a first step, we have converted easy controllers including
error page controllers.

For Scala controller, we have to define 3 methods:
  parseRequest
  executeAction
  renderResult

The first method is used to parse requests. The parsed params are propagated
to executeAction, which generates calculated values. The calculated values are
passed to renderResult, which renders the final results.

Related to #288