Ask before finishing modification. refs #418
to detect modification, I count number of input box in page.
if page contains visible and enabled input elements, it means
that user is still modifing some attributes of event.

to use button to judge, "save" button should be disabled by default.
I added some code to switch enabled/disabled status of "save" button.

I know that it is bad way for logic to use status of DOM,
but I think it is acceptable in this case: other way costs too much.